1st-part/functions/requirements.txt
# Function dependencies:

google-cloud-bigquery==1.24.0
google-cloud-pubsub==1.3.0
